+++ stm-flight-sw/limestone-firmware/Core/Src/logging.c
@@ -459,7 +459,7 @@ void send_flash_full() {
 // 0 success, 1 semaphore timeout, 2 flash full, 3 memory error
 int log_lmp_packet(uint8_t *buf, size_t buflen) {
 	uint8_t outbuf[MAX_TELEMETRY_B64_SIZE];
-	uint8_t outlen = base64_encode(buf, buflen, outbuf, MAX_TELEMETRY_B64_SIZE, 1);
+	size_t outlen = base64_encode(buf, buflen, outbuf, MAX_TELEMETRY_B64_SIZE, 1);
 	if(outlen > 0) {
 		outbuf[0] = FLASH_TELEM_MARK;
 		outbuf[outlen] = '\n';
